wip: add workload demand
This is the starting setup for our first algorithm! Right
now, we are defining jobs in the CRD, and the minicluster is
made interactive so we do not need to worry about adding random
sleeps or waits, etc. We also have validation and member type
support for each algorithm (starting just with minicluster). Where
I am at now is that we are instantiating the workload-demand algorithm
(it is an interface) and it is receiving input from the gRPC sidecar, and
then in the "MakeDecision" function we have our jobs spec (from the CRD)
and queue status. Next (after my run) I am going to work on this interaction!
We will want the algorithm to prepare a response payload that directs the
gRPC service to run all the lammps jobs, and then on the operator side
we will update the jobs spec to indicate they are run (decrementing the count)
and on the next iteration check the queue status again for size, etc and
adjust as needed. This is still early but very cool so far.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>