feat: add workload-demand algorithm
This is the skeleton of the workload-demand algorithm. Currently, we are
able to define jobs in a matrix, apply the ensemble.yaml, and then our
ensemble member (a minicluster) is created. The sidecar boots up
and starts the gRPC server, and the operator connects to it first
to get a status (the queue/jobs). At first hit there will be no
jobs or anything running. We then check the CRD status and see
there are jobs in the matrix. Since this algorithm is retroactive
(meaning it just submits a ton of work and then responds to it
with scaling and then termination) we submit all the jobs from
the matrix, and in the order they are supplied (or extended out).
Arguably this could be modified with an option for randomization,
by label, etc. The jobs are then submit on the MiniCluster, and
we check for subsequent queue status. This is up to where I have
implemented for this commit. Next I need to act on the MakeDecision
that the algorithm returns after receiving the queue, which will
typically be a request to scale up, down, or terminate. More
on this tomorrow.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>