feat: support for automated termination
This adds support for a member to terminate when we have
reached threshold, meaning that subsequent inactive statuses
have exceeded some threshold (defaults to 10, the example
has it set to 2, this is an algorithm option). I just need
to add the scaling logic here and this algorithm will be done!
But want to get some more sleep first

Signed-off-by: vsoch <vsoch@users.noreply.github.com>