fix: cooldown for scaling
Problem: we currently can trigger scaling events one after
the other because the reconcile happens immediately with the
update for the CRD, and then the queue status has not
changed (so it warrants another scaling event). With this
strategy, we add a new action for a resetCounter event, and
then reset both conters for scaling. This works beautifully
to have a scaling event, and then see the scaling counters
back at 0, and then more reasonably see the next scaling event
in 30 seconds to 1 minutes (unless the jobs finish and the
cluster is terminated first!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>