it's working!
This is so exciting! The ensemble runs, and I realize that we need
a setting in the yaml that is a termination signal - one
that says the library should exit. Right now it just keeps going :)
After that, this PR can merge for a very basic setup, and then I can
write the specific MiniCluster logic that will allow scaling. I will
say it again - this is really exciting.'

Signed-off-by: vsoch <vsoch@users.noreply.github.com>