wip: remove complexity (#17)
* wip: remove complexity

This is the start of larger changes to remove the complexity of this
operator. Instead of managing a python api here, we will use
ensemble-python that is able to orchestrate ensembles regardless
of the operator. The design is simple in that an ensemble is a
definition of jobs and rules, where each rule is a trigger and
action. Then the operator here simply needs to create the ensemble
member (e.g. the flux operator and there will be others) and there
is a deployment container service that is a main orchestrator
between ensemble members. The ensemble operator only needs to be
pinged by the deployment service when there is a request to scale
or similar. This design will also afford this deployment handling
orchestration between members, which the previous did not allow.
This is not fully working yet but I want to put the changes online
in case my computer poops.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* it's working!

This is so exciting! The ensemble runs, and I realize that we need
a setting in the yaml that is a termination signal - one
that says the library should exit. Right now it just keeps going :)
After that, this PR can merge for a very basic setup, and then I can
write the specific MiniCluster logic that will allow scaling. I will
say it again - this is really exciting.'

Signed-off-by: vsoch <vsoch@users.noreply.github.com>